AI-teams-controller-public/experiments/soniox/requirements.txt
soniox>=1.10.0
websockets>=12.0
python-dotenv>=1.0.0
pyaudio>=0.2.14
